Started (and nearly finished) adding trap encounter functionality
Added a Mechanical Skill to spawned heroes. It is reported in AnnounceHeroStats but doesn't do anything yet.
Added a Diplomacy Skill to spawned heroes. It is reported in AnnounceHeroStats but doesn't do anything yet.
Changed all instances of "Combat Power" to "Combat Skill". Hero skills are now on the same page, descriptively.
Changed skill displays to a percentage instead of xx/100 or "Combat Skill 70". Succinctness? We'll see.

Current task:
	Create at least one trap to disarm (or smash)
	Progress: Traps mostly work! Need to finish a reliable combat loop for smashing them (on purpose or out of necessity), then clean up stat display.